DetailFragment and HomeFragment done. Database handling issues.

+++ app/src/main/java/com/followapp/mytasks/common/entities/Task.kt
@@ -10,7 +10,7 @@ import java.util.Date
 data class Task(
     @ColumnInfo(name = "title") var title: String,
     @ColumnInfo(name = "is_done") var isDone: Boolean,
-    @PrimaryKey(autoGenerate = true) var id: Int = 0,
+    @PrimaryKey(autoGenerate = true) var id: Long = 0L,
     @ColumnInfo(name = "description") var description: String? = null,
     @ColumnInfo(name = "due_date") var dueDate: Date? = null,
 //    @ColumnInfo(name = "time_required") var timeRequired: LocalTime? = null,
